manually integrating interaction logic
brought in changes related to drawing two canvas layers. they're currently not positioning correctly. they should both be on top of the bingo-card div. the middle layer is going to be for drawing the lines indicating a bingo, but to prevent that layer from obstructing clickthrough, i've created a top canvas to handle click events and pass that through to the cells at the bottom. the main issue at this time is that the canvases are not rendering in the correct location. they will eventually not have borders, they're just bordered rn so i can see where they are